Fix mechanical signature mismatches + add known-alias filter
AchievementManager: unsigned long -> unsigned int (2 functions)
ItemManager::UnequipItem: unsigned int -> unsigned long (binary uses long)
BaseScreen::DrawBorders: SmartPtr const& -> by-value (match binary)
GameSound::SFXPlay: Delegate const& -> by-value
PowerUpManager::ActivatePower: Vec3* -> Vec3 by-value + call sites

check-signatures.py: add KNOWN_SIGNATURE_ALIASES table
  Draw(float*) <-> Draw(Vec3 const&, int)  (RE confirmed: HUD scale)
  PreDraw(float*) <-> PreDraw(Vec3 const&)  (same)
  Draw(float*) <-> Draw(Renderer&)          (port-specific)
  These are intentional port refactorings, not signature bugs.
  Report now shows filtered count separately.

+++ src/game/ItemManager.cpp
@@ -505,7 +505,7 @@ int ItemManager::EquipItem(unsigned int hash) {
 // Binary: lookup m_ByHash; if found, UnEquip() + SetEquippedItem(type, nullptr).
 // ASM-verified: 2026-05-23 binary @ 0x0010314c (re-analyst)
 // -----------------------------------------------------------------------
-bool ItemManager::UnequipItem(unsigned int hash) {
+bool ItemManager::UnequipItem(unsigned long hash) {
     std::map<uint32_t, ItemInfo*>::iterator it = m_ByHash.find((uint32_t)hash);
     if (it == m_ByHash.end()) return false;
     ItemInfo* item = it->second;
